# backend/docs/swagger.yaml
basePath: /api
definitions:
  auth.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  bet.PlaceBetInput:
    properties:
      amount:
        type: integer
      choice:
        type: string
      matchId:
        type: integer
    required:
    - amount
    - choice
    - matchId
    type: object
  match.CreateMatchInput:
    properties:
      startTime:
        type: string
      teamA:
        type: string
      teamB:
        type: string
    required:
    - startTime
    - teamA
    - teamB
    type: object
  match.ResolveMatchInput:
    properties:
      result:
        description: '''A'', ''B'''
        type: string
    required:
    - result
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  user.UpdateUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      points:
        type: integer
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 這是一個基於 DDD 架構開發的世足賽下注系統後端 API
  title: FIFA World Cup Betting System API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: User Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: Registration details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: User Registration
      tags:
      - Auth
  /bet:
    get:
      consumes:
      - application/json
      description: Returns all bets placed by the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user bets
      tags:
      - Bet
    post:
      consumes:
      - application/json
      description: User places a bet on a match (A/B)
      parameters:
      - description: Bet details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/bet.PlaceBetInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Place a bet
      tags:
      - Bet
  /match:
    get:
      consumes:
      - application/json
      description: Returns list of matches including betting pool stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all matches
      tags:
      - Match
    post:
      consumes:
      - application/json
      description: Admin creates a new match
      parameters:
      - description: Match details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/match.CreateMatchInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create match
      tags:
      - Match
  /match/{id}/resolve:
    put:
      consumes:
      - application/json
      description: Admin resolves a match result (A/B)
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      - description: Result A/B
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/match.ResolveMatchInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve match
      tags:
      - Match
  /match/sync:
    post:
      consumes:
      - application/json
      description: Admin manually updates matches from external API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Sync matches
      tags:
      - Match
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user details (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User
  /user/all:
    get:
      consumes:
      - application/json
      description: Returns list of all users, Admin only
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - User
  /user/daily:
    post:
      consumes:
      - application/json
      description: Allows user to claim their 1000 daily points
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Claim daily points
      tags:
      - User
  /user/leaderboard:
    get:
      consumes:
      - application/json
      description: Returns top 10 users by points
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Leaderboard
      tags:
      - User
  /user/me:
    get:
      consumes:
      - application/json
      description: Returns profile data of the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"